requirements.txt
fastapi>=0.100.0
uvicorn==0.21.1
requests==2.28.2
pydantic>=2.0.0
python-dotenv==1.0.0
python-anvil>=0.3.0
